+++ servers/quarkus-common/src/main/java/org/projectnessie/quarkus/config/QuarkusJdbcConfig.java
@@ -30,7 +30,8 @@ import org.projectnessie.versioned.storage.jdbc2.Jdbc2BackendBaseConfig;
  * the actual database to use. The property {@code nessie.version.store.persist.jdbc.datasource}
  * will be used to select one of the built-in datasources; currently supported values are: {@code
  * postgresql} (which activates the PostgresQL driver), {@code mariadb} (which activates the MariaDB
- * driver), and {@code mysql} (which targets MySQL backends, but using the MariaDB driver).
+ * driver), {@code mysql} (which targets MySQL backends, but using the MariaDB driver), {@code h2},
+ * and {@code mssql} (Microsoft SQL Server via the Microsoft JDBC driver).
  *
  * <p>For example, to configure a PostgresQL connection, the following configuration should be used:
  *
@@ -74,6 +75,18 @@ import org.projectnessie.versioned.storage.jdbc2.Jdbc2BackendBaseConfig;
  *   <li>{@code nessie.version.store.persist.jdbc.datasource=h2}
  * </ul>
  *
+ * <p>To connect to Microsoft SQL Server, the following configuration should be used:
+ *
+ * <ul>
+ *   <li>{@code nessie.version.store.type=JDBC2}
+ *   <li>{@code nessie.version.store.persist.jdbc.datasource=mssql}
+ *   <li>{@code
+ *       quarkus.datasource.mssql.jdbc.url=jdbc:sqlserver://localhost:1433;databaseName=my_database;encrypt=true}
+ *   <li>{@code quarkus.datasource.mssql.username=<your username>}
+ *   <li>{@code quarkus.datasource.mssql.password=<your password>}
+ *   <li>Other SQL Server-specific properties can be set using {@code quarkus.datasource.mssql.*}
+ * </ul>
+ *
  * Note: for MySQL, the MariaDB driver is used, as it is compatible with MySQL. You can use either
  * {@code jdbc:mysql} or {@code jdbc:mariadb} as the URL prefix.
  *
@@ -85,8 +98,8 @@ public interface QuarkusJdbcConfig extends Jdbc2BackendBaseConfig {
 
   /**
    * The name of the datasource to use. Must correspond to a configured datasource under {@code
-   * quarkus.datasource.<name>}. Supported values are: {@code postgresql} {@code mariadb}, {@code
-   * mysql} and {@code h2}. If not provided, the default Quarkus datasource, defined using the
+   * quarkus.datasource.<name>}. Supported values are: {@code postgresql}, {@code mariadb}, {@code
+   * mysql}, {@code h2}, and {@code mssql}. If not provided, the default Quarkus datasource, defined using the
    * {@code quarkus.datasource.*} configuration keys, will be used (the corresponding driver is
    * PostgresQL). Note that it is recommended to define "named" JDBC datasources, see <a
    * href="https://quarkus.io/guides/datasource#jdbc-configuration">Quarkus JDBC config