+++ servers/quarkus-common/src/main/java/org/projectnessie/quarkus/providers/storage/Jdbc2BackendBuilder.java
@@ -83,14 +83,15 @@ public class Jdbc2BackendBuilder implements BackendBuilder {
     String databaseKind = config.dbKind().get();
     if (!DatabaseKind.isPostgreSQL(databaseKind)
         && !DatabaseKind.isH2(databaseKind)
-        && !DatabaseKind.isMariaDB(databaseKind)) {
+        && !DatabaseKind.isMariaDB(databaseKind)
+        && !DatabaseKind.isMsSQL(databaseKind)) {
       throw new IllegalArgumentException(
           "Database kind for datasource "
               + dataSourceName
               + " is configured to '"
               + databaseKind
               + "', which Nessie does not support yet; "
-              + "currently PostgreSQL, H2, MariaDB (and MySQL via MariaDB driver) are supported. "
+              + "currently PostgreSQL, H2, MariaDB (and MySQL via MariaDB driver), and Microsoft SQL Server are supported. "
               + "Feel free to raise a pull request to support your database of choice.");
     }
   }